Use pattern matching to check for index when parsing `br_table` instruction
This is how we do it everywhere else, and there’s no reason for it to be
different here. The pattern match also has the advantage of tolerating
non-string values (e.g. arrays), which makes the #can_read_list? call
unnecessary.

Removing the numbered parameter changes the block’s arity, so we need to
add an explicit dummy parameter to keep #repeatedly happy. In hindsight
I don’t think the `until:` API is a good idea so I’m going to change
that soon.