Don’t treat empty block type as a type use
Per the spec [0], “the special case of a type use that is syntactically
empty […] is not regarded as an abbreviation for an inline function
type, but is parsed directly into an optional value type”.

The tests in this commit enforce the constraint that an empty block type
should be left alone by any desugaring of the type use abbreviation
since it’s not a type use. We don’t have any such desugaring yet, so the
tests don’t currently constrain anything, but they’ll prevent
regressions when we do implement it.

[0] https://webassembly.github.io/spec/core/text/instructions.html#text-blocktype