Return deferred result from #process_typeuse
To desugar the type use abbreviation [0] we’re going to need to know the
complete list of type definitions for the current module, but that list
won’t become available until after all of the module’s fields have been
processed.

So we have to return some value from #process_typeuse to keep
preprocessing moving, but we won’t be ready to compute the final
desugared S-expression until later. To handle that, we’ll make
the #process_typeuse method defer its result by returning a proc which
can be called at a future time when the type definitions are available.
The #after_all_fields helper makes this a little tidier by having a
meaningful name and allowing its block to accept zero arguments if it
doesn’t have any use for the type definitions.

It’s important that we consume all of the relevant input before we
create the proc because the immediate side-effect of #process_typeuse on
the current S-expression is an essential part of its behaviour. So the
call to #after_all_fields only goes around the (side-effect–free)
construction of the result array, not the calls to #process_parameters
and #process_results.

The #process_typeuse method has four callers:

  * #process_function_definition
  * #process_instructions
  * #process_blocktype
  * #process_import_descriptor

These callers don’t have access to the list of type definitions
that #process_typeuse needs, but they still need to return a result
which contains a fully processed type use, so they call the proc with a
dummy empty array for now. And because that dummy array’s not useful, we
don’t actually use the type definitions in #process_typeuse yet.

Having made this change, we next want to push the deferred result
further up the call graph until it reaches #process_module, where we
eventually expect to be able to provide it with a meaningful list of
type definitions which we can properly use in #process_typeuse. By
choosing a sensible order in which to convert each caller to return a
deferred result we can minimise the amount of churn and confusion
introduced by the intermediate refactoring steps.

Starting from #process_module, the static call graph of all methods
which transitively call #process_typeuse looks like this:

  #process_module
  └─ #process_fields
     └─ #process_field
        ├─ #process_function_definition
        │  ├─ #expand_inline_import_export
        │  │  ├─ #expand_inline_import
        │  │  │  └─ (#process_fields…)
        │  │  └─ #expand_inline_export
        │  │     └─ (#process_fields…)
        │  ├─ (#process_typeuse…)
        │  └─ (#process_instructions…)
        ├─ #process_table_definition
        │  ├─ #expand_inline_element_segment
        │  │  └─ (#process_fields…)
        │  └─ (#expand_inline_import_export…)
        ├─ #process_memory_definition
        │  ├─ #expand_inline_data_segment
        │  │  └─ (#process_fields…)
        │  └─ (#expand_inline_import_export…)
        ├─ #process_global_definition
        │  ├─ (#expand_inline_import_export…)
        │  └─ (#process_instructions…)
        ├─ #process_import
        │  └─ #process_import_descriptor
        │     └─ (#process_typeuse…)
        ├─ #process_element_segment
        │  ├─ #process_active_element_segment
        │  │  ├─ #process_offset
        │  │  │  ├─ (#process_instructions…)
        │  │  │  └─ (#process_instruction…)
        │  │  └─ #process_element_list
        │  │     └─ #process_element_expressions
        │  │        └─ #process_element_expression
        │  │           ├─ (#process_instructions…)
        │  │           └─ #process_instruction
        │  │              └─ #process_instructions
        │  │                 ├─ (#process_typeuse…)
        │  │                 ├─ (#process_instructions…)
        │  │                 └─ #process_blocktype
        │  │                    └─ #process_typeuse
        │  ├─ #process_declarative_element_segment
        │  │  └─ (#process_element_list…)
        │  └─ #process_passive_element_segment
        │     └─ (#process_element_list…)
        └─ #process_data_segment
           └─ #process_active_data_segment
              └─ (#process_offset…)

Ignoring recursive calls to #process_fields and #process_instructions,
each method is shown primarily at its maximum depth in the call graph.
This allows us to organise the methods into distinct groups whose
members all share the same maximum depth:

  * #process_typeuse (deepest)
  * #process_blocktype
  * #process_instructions
  * #process_instruction
  * #process_element_expression
  * #process_element_expressions
  * #expand_inline_import, #expand_inline_export, #process_offset and
    #process_element_list
  * #expand_inline_element_segment, #expand_inline_data_segment,
    #expand_inline_import_export, #process_import_descriptor,
    #process_active_element_segment, #process_declarative_element_segment,
    #process_passive_element_segment and #process_active_data_segment
  * #process_function_definition, #process_table_definition,
    #process_memory_definition, #process_global_definition,
    #process_import, #process_element_segment and #process_data_segment
  * #process_field
  * #process_fields
  * #process_module (shallowest)

By starting with the deepest method and working our way up to the
shallowest, we’ll ensure that all of a method’s callees will have been
converted to return a deferred result by the time we try to convert the
method itself, which allows us to make progress continuously upwards in
the call graph without needing to backtrack.

In the following commits we’ll work through the methods in this order
until the deferred result has reached #process_module.

[0] https://webassembly.github.io/spec/core/text/modules.html#abbreviations