Preprocess block type for `block`/`loop`/`if` instructions
At the moment we’re treating any occurrence of `result` as a type use,
but that brute force approach isn’t correct because `result` can appear
in a block type [0] without necessarily being a type use, so we need to
be a bit more surgical.

d0f567570c3740ed84c9da881cf1652a6ccac047 already added test coverage for
this behaviour when it introduced the initial “hack” to expand parameter
and result abbreviations inside instructions, so we don’t need to test
it again here.

We’re not yet observing the rules about whether a block type constitutes
a type use, but introducing #process_blocktype gives us a place to do it
in future.

[0] https://webassembly.github.io/spec/core/text/instructions.html#text-blocktype