Update preprocessor tests to not depend upon missing `call_indirect` index abbreviation
Because a `call_indirect` instruction may contain `param` and `result`
occurrences which need to be desugared, it provides an easy way for a
test to demonstrate that a particular place in an S-expression is being
preprocessed successfully.

Consequently quite a few of the preprocessor tests use `call_indirect`,
specifically its abbreviated syntax which allows an explicit table index
[0] to be omitted and automatically replaced with zero when
abbreviations are expanded. To prevent these tests from breaking when we
implement desugaring of the missing `call_indirect` index abbreviation
we need to update them to contain explicit table indexes which are
unaffected by the expansion of that abbreviation.

[0] https://webassembly.github.io/spec/core/text/instructions.html#abbreviations