Update preprocessor tests to not depend upon folded instruction abbreviation
Most of the preprocessor tests use the convenient folded instruction
syntax [0] and assume it’ll be unchanged by preprocessing. To prevent
these tests from breaking when we implement desugaring of the folded
instruction abbreviation we need to update them to use unfolded
instructions which are unaffected by the expansion of that abbreviation.

[0] https://webassembly.github.io/spec/core/text/instructions.html#text-foldedinstr