Read complete table instructions
Different table instructions accept different numbers of indexes as
immediate arguments [0], some of which are optional [1]. For the
purposes of reading a complete instruction, the easiest thing is just to
read as many indexes as possible and let another part of the parser
worry about whether that’s correct.

Here we’re opportunistically lumping in `br_table` with the actual table
instructions; although it takes label indexes [2] rather than table
indexes as immediate arguments, that distinction doesn’t matter
lexically.

[0] https://webassembly.github.io/spec/core/text/instructions.html#table-instructions
[1] https://webassembly.github.io/spec/core/text/instructions.html#id1
[2] https://webassembly.github.io/spec/core/text/instructions.html#text-plaininstr