Process complete `block` and `loop` in #process_structured_instruction
We were previously stopping after the block type and letting repeated
calls to #process_instruction mop up the rest (including the `end`
keyword), but in the case of `block` and `loop` we’re now reading the
entire instruction at once and processing its parts appropriately.

In practice it’s important that the end label (if present) matches the
opening label, but our job is only to expand abbreviations, so we’ll
ignore that detail and let the parser take responsibility for checking
syntactic validity later.