Read complete `load` and `store` instructions
The various `load` and `store` instructions [0] take two optional
immediate arguments:

  `i32.load` (`offset=` u32)? (`align=` u32)?
  `i64.load` (`offset=` u32)? (`align=` u32)?
  `f32.load` (`offset=` u32)? (`align=` u32)?
  `f64.load` (`offset=` u32)? (`align=` u32)?
  `i32.load8_s` (`offset=` u32)? (`align=` u32)?
  `i32.load8_u` (`offset=` u32)? (`align=` u32)?
  ⋮
  `i32.store` (`offset=` u32)? (`align=` u32)?
  `i64.store` (`offset=` u32)? (`align=` u32)?
  `f32.store` (`offset=` u32)? (`align=` u32)?
  `f64.store` (`offset=` u32)? (`align=` u32)?
  `i32.store8` (`offset=` u32)? (`align=` u32)?
  ⋮

To read a complete `load` or `store` instruction we need to read these
arguments if present and return them verbatim.

To keep the regular expression concise I’ve assumed that any keyword
which begins “i32.load…” or “i32.store…” (likewise for “i64”, “f32” and
“f64”) indicates a `load` or `store` instruction, which may turn out to
be incorrect on an infinite timescale but is at least true for the
current version of WebAssembly.

[0] https://webassembly.github.io/spec/core/text/instructions.html#memory-instructions