Desugar folded `if` abbreviation
Instead of processing a folded `if` directly, we desugar it by
rearranging its pieces [0] into its constituent instructions and then
recursively desugaring those by calling #process_instructions, in the
same spirit as #expand_inline_element_segment, #expand_inline_import
etc.

It’s important that we desugar folded `if` occurrences before (versus
after) we desugar folded `block`/`loop` occurrences, because desugaring
a folded `block` or `loop` in the condition of a folded `if` would
result in a syntactically invalid folded `if` — its condition would no
longer be a sequence of folded instructions, which the grammar says it
must be. By doing it in this order, we remove any folded `if`
instructions from the preprocessed output and thereby no longer require
the parser to deal with the syntax of folded `if` at all, which gives us
the freedom to safely desugar `block`/`loop` next.

[0] https://webassembly.github.io/spec/core/text/instructions.html#text-foldedinstr