Replace unhelpful named block parameters with numbered ones in preprocessor
In all of these cases it’s either obvious or uninteresting what kind of
value is being passed to the block, so we shouldn’t draw attention to it
with either a repetitive name or a meaningless one like `result`. By
using a numbered parameter we make the code more concise and draw
attention to the block’s body rather than its argument.