Push #read_instructions into #parse_structured_instruction cases in AST parser
Crucially, this removes the expectation that #read_instructions will
ever consume the entire body of an `if` (including the `else`), which
will in turn allow us to make #read_instructions unconditionally stop on
`else` in a future commit. And that’s the behaviour we want, because
`else` is not an instruction.